backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
django-environ>=0.11.0
django-filter>=23.5
drf-spectacular>=0.27.0

# Database
psycopg2-binary>=2.9.9

# Async / WebSocket
channels>=4.0.0
daphne>=4.0.0
channels-redis>=4.1.0

# Celery
Celery>=5.3.0
redis>=5.0.0
django-celery-beat>=2.5.0

# Payments
razorpay>=1.4.0

# Firebase (push notifications)
firebase-admin>=6.3.0

# Storage
boto3>=1.34.0
django-storages>=1.14.0

# Validation
phonenumbers>=8.13.0
django-phonenumber-field[phonenumberslite]>=7.3.0

# QR generation
qrcode>=7.4.0
Pillow>=10.0.0

# Security
cryptography>=41.0.0

# Production servers
gunicorn>=21.2.0
uvicorn[standard]>=0.24.0
whitenoise>=6.6.0
